Update patch set 2
Patch Set 2:

Is the plan to upstream most of these fixes? I am a bit concerned about the _WIN64 ifdef check, mostly because that may have some different impact on MSVC types, but the rest of it is pretty straightforward (and a huge improvement). Thanks for working on this.

Patch-set: 2
CC: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>