Differentiate XFAIL and XFAIL-BUILD for run tests.
When test runs are deferred, we need to know if we expect the test to
fail at runtime (in which case xfail handling is deferred to the
bundler) or a build time (in which case we'll never make it to the
bundler). This adds XFAIL-BUILD to indicate that for .pass.cpp tests.
Other forms of tests are build-only so those continue using normal
XFAIL.

This commit corrects any annotations that were failing on my Linux
machine, but there isn't a good way to find these for existing tests
other than trial and error.

Change-Id: I3a66584d79a12c717eccd3a1f7c8c45d6b3ce222